feat: Add skip link to skip header and google map
Provide a way for users of assistive technology to skip 
a header that is not part of a site's essential content and might also 
have a long navigation (many links).
Allow users to skip google maps so that they have not to tab
through all the elements of the map. Important info, such as 
restaurant's address, can be found in restaurant's details section.